Rename netatmo data_handler to coordinator
