+++ homeassistant/components/netatmo/coordinator.py
