+++ tests/components/netatmo/test_camera.py
@@ -474,7 +474,7 @@ async def test_camera_reconnect_webhook(
         patch(
             "homeassistant.components.netatmo.api.AsyncConfigEntryNetatmoAuth"
         ) as mock_auth,
-        patch("homeassistant.components.netatmo.data_handler.PLATFORMS", ["camera"]),
+        patch("homeassistant.components.netatmo.coordinator.PLATFORMS", ["camera"]),
         patch(
             "homeassistant.components.netatmo.async_get_config_entry_implementation",
         ),
@@ -587,7 +587,7 @@ async def test_camera_webhook_consistency(
         patch(
             "homeassistant.components.netatmo.api.AsyncConfigEntryNetatmoAuth"
         ) as mock_auth,
-        patch("homeassistant.components.netatmo.data_handler.PLATFORMS", ["camera"]),
+        patch("homeassistant.components.netatmo.coordinator.PLATFORMS", ["camera"]),
         patch(
             "homeassistant.components.netatmo.async_get_config_entry_implementation",
         ),
@@ -707,7 +707,7 @@ async def test_setup_component_no_devices(
         patch(
             "homeassistant.components.netatmo.api.AsyncConfigEntryNetatmoAuth"
         ) as mock_auth,
-        patch("homeassistant.components.netatmo.data_handler.PLATFORMS", ["camera"]),
+        patch("homeassistant.components.netatmo.coordinator.PLATFORMS", ["camera"]),
         patch(
             "homeassistant.components.netatmo.async_get_config_entry_implementation",
         ),
@@ -750,7 +750,7 @@ async def test_camera_image_raises_exception(
         patch(
             "homeassistant.components.netatmo.api.AsyncConfigEntryNetatmoAuth"
         ) as mock_auth,
-        patch("homeassistant.components.netatmo.data_handler.PLATFORMS", ["camera"]),
+        patch("homeassistant.components.netatmo.coordinator.PLATFORMS", ["camera"]),
         patch(
             "homeassistant.components.netatmo.async_get_config_entry_implementation",
         ),
