+++ tests/components/netatmo/test_light.py
@@ -127,7 +127,7 @@ async def test_setup_component_no_devices(hass: HomeAssistant, config_entry) -> 
         patch(
             "homeassistant.components.netatmo.api.AsyncConfigEntryNetatmoAuth"
         ) as mock_auth,
-        patch("homeassistant.components.netatmo.data_handler.PLATFORMS", ["light"]),
+        patch("homeassistant.components.netatmo.coordinator.PLATFORMS", ["light"]),
         patch(
             "homeassistant.components.netatmo.async_get_config_entry_implementation",
         ),
